finx-agentic/pyproject.toml
[project]
name = "finx-agentic"
version = "0.1.0"
description = "A modular AI agent system built with the Agno framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=0.1.0",
    "google-genai>=1.60.0",
    "google-generativeai>=0.8.6",
    "mcp>=1.26.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.2",
    "jinja2>=3.1.0",
    "boto3>=1.34.0",
    "atlassian-python-api>=3.41.0",
    "neo4j>=5.24.0",
    "numpy>=1.26.0",
    "litellm>=1.0.0",
    "graphiti-core[falkordb]>=0.5.0",
    "openai>=1.0.0",
    "langtrace-python-sdk>=3.0.0",
    "agentops>=0.4.21",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.1.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
